requirements.txt
mutagen
librosa
numpy
tqdm
fastapi
spotipy
pandas
python-dotenv
requests
